hw2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
albumentations>=1.0.0
timm>=0.9.0 
torchmetrics>=1.0.0
torchvision>=0.15.0
